automake: Add full dependencies for ddd.info.txt.gz.C to dist
The problem is that during distcheck it will try to rebuild
ddd.info.txt.gz.C because the dependencies are not there. Unfortunately,
since the source tree is made to be read-only when testing VPATH builds
and ddd.info.txt.gz.C's prerequisities require manipulating the source
tree, we fail here. So we provide all dependencies.